+++ CHANGELOG.md
@@ -5,9 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [###VERSION###] - 2019-07-03
-### Added
+### Fixed
+- NXMConversationEventsController returns the last event
 - CallMember status calculated by the current leg status.
 
+### Added
 Added initiators property to NXMMember 
 ```
 @property (nonatomic, readonly, nonnull) NSDictionary<NSValue *, NXMInitiator *> *initiators;